Cleaned, fixed, commented
i cleaned the code and removed uneccesary parts, and replaced the few reppitions.
i commented the code to make it more understandable - will be more important when i do the code for the actual model.
fixed the minor bugs, a main bug was the partioning of data, it used the nodes instead of the amount of data (because of the parameters i had set i didn't notice)
i sorted the files to make it more organised - total waste of time!!!